refactor: simplify scripts
