--- scripts/index.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-require("@swc-node/register")
-require("./src/index.ts")
