+++ scripts/logger.ts
@@ -1,0 +1,5 @@
+import signale from "signale"
+
+export function createLogger(scope: string) {
+  return signale.scope(scope)
+}
