--- scripts/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@zag-js/cli",
-  "version": "1.0.0",
-  "private": true,
-  "bin": {
-    "zag": "index.js"
-  },
-  "devDependencies": {
-    "@babel/generator": "7.18.7",
-    "@babel/parser": "7.18.8",
-    "@babel/traverse": "7.18.8",
-    "@babel/types": "7.18.8"
-  },
-  "dependencies": {
-    "@manypkg/get-packages": "1.1.3",
-    "@lerna/package-graph": "5.1.6",
-    "@swc-node/register": "1.5.1",
-    "@types/signale": "1.4.4",
-    "commander": "9.3.0",
-    "relative": "3.0.2",
-    "signale": "1.4.0"
-  }
-}
