--- scripts/src/relative.d.ts
@@ -1,3 +1,0 @@
-declare module "relative" {
-  export default function relative(from: string, to: string): string
-}
