--- scripts/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "esModuleInterop": true,
-    "downlevelIteration": true
-  },
-  "include": ["src"]
-}
